+++ lib/compony/components/edit.rb
@@ -29,16 +29,14 @@ module Compony
 
         label(:long) { |data| I18n.t('compony.components.edit.label.long', data_label: data.label) }
         label(:short) { |_| I18n.t('compony.components.edit.label.short') }
-        icon { :pencil }
 
         form_cancancan_action :edit
 
         exposed_intents do
           if data_class.owner_model_attr
             add :show, @data.send(data_class.owner_model_attr),
-                button: { icon: :xmark, color: :secondary },
-                label:  I18n.t('compony.cancel'),
-                name:   :back_to_owner
+                label: I18n.t('compony.cancel'),
+                name:  :back_to_owner
           end
         end