+++ lib/compony/components/show.rb
@@ -14,14 +14,12 @@ module Compony
 
         label(:long) { |data| data.label } # rubocop:disable Style/SymbolProc
         label(:short) { |_| I18n.t('compony.components.show.label.short') }
-        icon { :eye }
 
         exposed_intents do
           if data_class.owner_model_attr
             add :show, @data.send(data_class.owner_model_attr),
-                button: { icon: :xmark, color: :secondary },
-                label:  I18n.t('compony.back'),
-                name:   :back_to_owner
+                label: I18n.t('compony.back'),
+                name:  :back_to_owner
           end
 
           if Compony.comp_class_for(:edit, family_name)